Changes for group information allocation and displaying fuel
First responses don't get assigned to group 0 - had to edit code to account for this so group info went to the correct group and the rockets were placed in the correct places.
Also, beginning of displaying and calculating of the fuel. Just need to prevent the rockets from flying if fuel is not full